refactored hb tasks to directly write files to variation folder
