+++ gulp/tasks/hb.js
@@ -1,59 +1,16 @@
 const config = require('./../config');
 const gulp = require('gulp');
-const path = require('path');
-const replace = require('gulp-replace');
-const rename = require('gulp-rename');
-const notify = require("gulp-notify");
 const hb = require('gulp-hb');
 const watch = require('gulp-watch');
 const runSequence = require('run-sequence');
+const hbHelper = require('../lib/handlebars-helper');
 
 gulp.task('hb', function () {
-    const hbStream = hb({ debug: false })
-        .partials(path.join(config.cwd, 'src/layouts/*.hbs'))
-        .partials(path.join(config.cwd, 'src/components/*/partials/**/*.hbs'));
-
-	return gulp
-		.src(config.srcDir + '/components/*/variations/**/index.hbs')
-		.pipe(hbStream)
-		.on('error', notify.onError(function (error) {
-			return {
-				title: 'hb',
-				message: error.message
-			};
-		}))
-		.pipe(rename({extname: ".html"}))
-        .pipe(rename((path) => {
-            path.dirname = path.dirname.replace('/variations', '');
-        }))
-		.pipe(gulp.dest(config.devDir));
+    return hbHelper.hbTask(config.devDir, false);
 });
 
 gulp.task('hb:dist', function () {
-    const hbStream = hb({ debug: false })
-        .partials(path.join(config.cwd, 'src/layouts/*.hbs'))
-        .partials(path.join(config.cwd, 'src/components/*/partials/**/*.hbs'))
-        .data({ isProduction: true });
-
-    let stream = gulp
-        .src(config.srcDir + '/components/*/variations/**/index.hbs')
-        .pipe(hbStream)
-        .on('error', notify.onError(function (error) {
-            return {
-                title: 'hb',
-                message: error.message
-            };
-        }))
-        .pipe(rename({extname: ".html"}))
-        .pipe(rename((path) => {
-            path.dirname = path.dirname.replace('/variations', '');
-        }));
-
-    config.replaceStrings.forEach((replaceObject) => {
-       stream.pipe(replace(replaceObject.subStr, replaceObject.newSubStr))
-    });
-
-    return stream.pipe(gulp.dest(config.distDir));
+    return hbHelper.hbTask(config.distDir, true);
 });
 
 gulp.task('watch:hb', function () {
@@ -62,27 +19,9 @@ gulp.task('watch:hb', function () {
             ['hb']
         );
     });
-
 });
 
 
-
-// gulp.task('watch:static:hb', function () {
-// 	let files = [config.global.src + '/partials/**/*.hbs'];
-//
-// 	config.global.components.forEach(function(currentComponent) {
-// 		files.push(config.global.src + currentComponent +'/**/*.hbs');
-// 	});
-//
-// 	watch(files, config.watch, function () {
-// 		runSequence(
-// 			['static:hb']
-// 		);
-// 	});
-//
-// });
-//
-//
 // /**
 //  * indexr creates the preview file index
 //  */
