+++ src/components/aem-module-reloaded/variations/README.md
@@ -1,0 +1,4 @@
+# Component variations
+
+All variations have their own folder, normally the language, like "de", "en" and so on
+The variation folder containes only a single file named "index.hbs"!!! All other files are ignored!
