+++ src/layouts/README.md
@@ -1,0 +1,5 @@
+# Layout templates
+
+Layout templates are only used for preview pages. They will not be copied to distribution folders.
+
+Important handlebars flag is `{{#unless isProduction}}`. This can be used to scope the markup that will be removed for the production templates
